ARROW-17717: [R] Lintr error on CI (#14113)
The new release of lintr has a new feature - giving a warning when code is commented out - which is causing the lintr-containing CI job to fail.  We could have pinned to a previous version of lintr, but that seem like a useful feature & the lines of code which trigger it seem like reasonable exceptions, so I've just added a #nolint flag on each of them.

Authored-by: Nic Crane <thisisnic@gmail.com>
Signed-off-by: Nic Crane <thisisnic@gmail.com>