ARROW-17736: [C++] Added a fallback name resolution mechanism to the Substrait producer. (#14143)
This will only be used if the URI part of the function reference is empty or /.  This is for compatibility with Isthmus which still hasn't decided what to use for the URI portion.  Longer term we may remove this.  Or we may leave it in as a utility for developers of new functions that don't want to have to worry about the URI initially.

Authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>