ARROW-17885: [R] Return BLOB data as list of raw instead of a list of integers (#14277)
This PR adds support for `blob::blob()`, which is common in R database land to denote "binary", and `vctrs::list_of()`, which is similar, easy, and helps a bit with list of things that happen to be all NULL.

We have our own infrastructure for binary and lists of things too, which I assume pre-dates the mature vctrs and blob? Should we consider having `as.vector()` output those objects instead of the custom `arrow_list/large_list/binary` classes we implement here?

Lead-authored-by: Dewey Dunnington <dewey@fishandwhistle.net>
Co-authored-by: Dewey Dunnington <dewey@voltrondata.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>