PARQUET-2188: [parquet-cpp] Add SkipRecords API to RecordReader (#14142)
The RecordReader is missing an API to skip records. There is a Skip method in the ColumnReader, but that skips based on the number of values/levels and not records. For repeated fields, this SkipRecords API will detect the record boundaries and correctly skip the right number of values for the requested number of records.

Lead-authored-by: Fatemah Panahi <panahi@google.com>
Co-authored-by: Fatemah Panahi <fatemehp@users.noreply.github.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Co-authored-by: fatemehp <fatemehp@users.noreply.github.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>