ARROW-18164: [Python] Honor default memory pool in Dataset scanning (#14516)
Dataset scanning was using the C++ default memory backend, and not the one set from python (if not specified explicitly). `maybe_unbox_memory_pool(memory_pool)` is meant to always be called (not only if `memory_pool is not None`), since this will then return the python default memory pool if `memory_pool` is not specified. 

Authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>