GH-33621: [Documentation][Developer Tools] Add CODEOWNERS file (#33622)
# Rationale for this change
CODEOWNERS will lift the burden of finding an initial reviewer for a PR from the contributor and enable committers to basically "subscribe" to a selection of PRs based on their interests/competence within the monorepo without having to watch _all_ notifications for the repo.

# What changes are included in this PR?

I have added a rough structure for the file and added myself to a selection of the patterns. Any committer please feel free to add yourself via suggestion or direct push to this branch.

* Closes: #33621

Lead-authored-by: Jacob Wujciak-Jens <jacob@wujciak.de>
Co-authored-by: David Li <li.davidm96@gmail.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Co-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Will Jones <willjones127@gmail.com>
Co-authored-by: Weston Pace <weston.pace@gmail.com>
Co-authored-by: Dewey Dunnington <dewey@dunnington.ca>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>