GH-33779: [R] Nightly builds (R 3.5 and 3.6) failing due to field refs test (#33780)
This PR fixes a test which is failing on R versions < 4.0 due to the use of the function `data.frame()` which had a change in default parameter value for `stringsAsFactors` (from `TRUE` to `FALSE`) in 4.0, and so is behaving differently in these versions.  

I've just swapped out the use of `data.frame()` for `tibble::tibble()` , which we use more prevalently in our tests, and has no such issues on earlier R version.

* Closes: #33779

Authored-by: Nic Crane <thisisnic@gmail.com>
Signed-off-by: Nic Crane <thisisnic@gmail.com>