GH-33760: [R][C++] Handle nested field refs in scanner  (#33770)
### Rationale for this change

Followup to https://github.com/apache/arrow/pull/19706/files#r1073391100 with the goal of deleting and simplifying some code. As it turned out, it was more about moving code from the R bindings to the C++ library.

### Are there any user-facing changes?

Not for R users, but this fixes a bug in the dataset C++ library where nested field refs could not be handled by the scanner.

* Closes: #33760

Authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>