GH-33864: [Go] Don't directly coerce cgo.Handle to unsafe.Pointer (#33865)
### Rationale for this change

Converting cgo.Handle to unsafe.Pointer is not allowed as per the docs. Doing so appears to cause occasional panics at runtime, possibly because something corrupts or collects the handle unexpectedly.

### What changes are included in this PR?

Heap-allocate the Handle to preserve it.

### Are these changes tested?

We were unable to reproduce this outside of an integration test.

### Are there any user-facing changes?

No user-facing changes.
* Closes: #33864

Authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: Matt Topol <zotthewizard@gmail.com>