GH-32050: [C++] Implement Rank kernel on chunked arrays (#33846)
This is a duplicate/continuation of https://github.com/apache/arrow/pull/14505.

William Ayd did the bulk of the work here. There was a somewhat nontrivial rebase that needed to occur (due to one of my recent PRs) so Will's commits have been squashed with the credits intact.

Besides that, I did some refactoring to avoid duplicating `ChunkedArraySorter`'s logic in `ChunkedArrayRanker`. 

* Closes: #32050

Lead-authored-by: benibus <bpharks@gmx.com>
Co-authored-by: Will Ayd <william.ayd@icloud.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>