MINOR: [Go] Use shared cache in SQLite example. (#33983)
### Rationale for this change

When calling this server with a few concurrent requests, you can easily force the Go stdlib to spawn a new connection. Since only the `:memory:` DSN is specified, the stdlib doesn't know any better and effectively creates a new database. The result is a lot of `no such table` errors.

### What changes are included in this PR?

This change adds the `?cache=shared` as specified at:

https://www.sqlite.org/sharedcache.html#enabling_shared_cache_mode

This ensures that these new connections participate in same in-memory database and can see all the tables created at server start.

### Are these changes tested?

I haven't written any additional tests here, because this behavior is difficult to test. Suggestions are welcome on an approach here.

### Are there any user-facing changes?

No.

* Closes: #33983

Authored-by: Brett Buddin <brett@buddin.org>
Signed-off-by: Matt Topol <zotthewizard@gmail.com>