GH-33904: [R] improve behavior of s3_bucket - work-around (#34009)
I think this is another option to address #33918.

This retains the existing behavior when no additional arguments (region, endpoint, etc) are supplied.  

If any optional arguments are supplied, this approach will by-pass the `$from_uri` call.  I think this is reasonable, given that the `from_uri()` call is implicitly ignoring optional arguments already, which is giving rise to the current bugs.  

* Closes: #33904

Lead-authored-by: Carl Boettiger <cboettig@gmail.com>
Co-authored-by: Carl Boettiger <cobettig@gmail.com>
Co-authored-by: Carl <cboettig@gmail.com>
Co-authored-by: Dewey Dunnington <dewey@voltrondata.com>
Signed-off-by: Dewey Dunnington <dewey@fishandwhistle.net>