GH-33143: [C++] Naming and doc/test changes for local_time compute kernel (#34263)
### Rationale for this change

A better naming for local_time kernel was proposed in post merge review of #34208.

### What changes are included in this PR?

Change `local_time` to `local_timestamp` and related docs/test changes.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

Changing `local_time` to `local_timestamp` is a user facing change. But since it was not yet released we can probably treat it as non-breaking.
* Closes: #33143

Authored-by: Rok Mihevc <rok@mihevc.org>
Signed-off-by: Rok Mihevc <rok@mihevc.org>